Relocated fetchAllData() fn to apiCalls.js
